Was added hooks to search but it's not working
